Separated seq and concurrent executions to facilitate better benchmarking
